Rename organization objects to match naming convention
JIRA: RAIL-2999
